Gamble quarter (And stats)

+++ src/main/java/net/kodehawa/mantarobot/commands/currency/CurrencyManager.java
@@ -3,9 +3,11 @@ package net.kodehawa.mantarobot.commands.currency;
 import net.kodehawa.mantarobot.data.MantaroData;
 
 public class CurrencyManager {
+	//Updated Jan 17
 	public static final long USD_ESTIMATIVE = 1500000000000L;
+
 	public static double creditsWorth() {
-		return USD_ESTIMATIVE / MantaroData.getData().get().users.values().stream()
+		return (double)USD_ESTIMATIVE / (double)MantaroData.getData().get().users.values().stream()
 			.mapToLong(userData -> userData.money).sum();
 	}
 }
